collab: Pass down staff usage in `UpdatePlan` message
